Rework languages module; /filters command to see filters instead of /getfilter
It now load languages from json file.
For add new language simply add new language in 'langs' folder, no need edit modules.

+++ sophie_bot/modules/language.py
@@ -1,23 +1,26 @@
 import re
+import ujson
+import os
+
 from telethon.tl.custom import Button
 from telethon import events
 
 from sophie_bot.events import register, flood_limit
 from sophie_bot.modules.users import is_user_admin
-from sophie_bot import REDIS, MONGO, bot
+from sophie_bot import REDIS, MONGO, LOGGER, bot
 
-from sophie_bot.modules.lang import EN_STRINGS, RU_STRINGS
+LANGUAGES = {}
+LANGS = ()
 
-SUPPORTED_LANGUAGES = {
-    'en': 'English 🇺🇸',
-    'ru': 'Russian 🇷🇺',
-}
+for filename in os.listdir('sophie_bot/modules/langs'):
+    f = open('sophie_bot/modules/langs/' + filename, "r")
+    lang = ujson.load(f)
+    exec("LANGUAGES[\"" + lang['language_info']['code'] + "\"] = lang")
+    LANGS += tuple([lang['language_info']['code']])
 
-LANG_VARS = {
-    'en': EN_STRINGS,
-    'ru': RU_STRINGS,
-}
+print(LANGUAGES)
 
+LOGGER.info("Languages loaded: {}".format(LANGS))
 
 @register(incoming=True, pattern="^/lang$")
 async def handler(event):
@@ -75,22 +78,32 @@ Please wait 3 minutes before using this command')
 @bot.on(events.CallbackQuery(data=re.compile(b'select_lang_')))
 async def event(event):
     chat = event.chat_id
+    K = await is_user_admin(chat, event.original_update.user_id)
+    if K is False:
+        await event.answer("You don't have rights to set language here!", alert=True)
+        return
     event_data = re.search(r'select_lang_(.*)', str(event.data))
     lang = event_data.group(1)[:-1]
     REDIS.set('lang_cache_{}'.format(chat), lang)
     old = MONGO.lang.find_one({'chat_id': chat})
-    if old:
+    if old: 
         MONGO.notes.delete_one({'_id': old['_id']})
     MONGO.lang.insert_one({'chat_id': chat, 'lang': lang})
     await event.edit(
-        "Language changed to **{}**!".format(SUPPORTED_LANGUAGES[lang]))
+        "Language changed to **{}**!".format(
+            LANGUAGES[lang]["language_info"]["name"] + \
+                " " + LANGUAGES[lang]["language_info"]["flag"]
+        ))
 
 
-def get_string(text, chat_id):
+def get_string(module, text, chat_id):
     lang = get_chat_lang(chat_id)
-    for H in LANG_VARS:
-        if H == lang and text in LANG_VARS[H]:
-            return LANG_VARS[H][text]
+    print(LANGUAGES[lang])
+    if not module in LANGUAGES[lang]['STRINGS']:
+        if text in LANGUAGES['en']['STRINGS'][module]:
+            return LANGUAGES['en']['STRINGS'][module][text]
+    if text in LANGUAGES[lang]['STRINGS'][module]:
+        return LANGUAGES[lang]['STRINGS'][module][text]
 
     return text
 
@@ -106,7 +119,7 @@ def get_chat_lang(chat_id):
             REDIS.set('lang_cache_{}'.format(chat_id), db_lang['lang'])
             return db_lang['lang']
         user_lang = MONGO.user_list.find_one({'user_id': chat_id})
-        if user_lang and user_lang['user_lang'] in SUPPORTED_LANGUAGES:
+        if user_lang and user_lang['user_lang'] in LANGS:
             # Add telegram language in lang cache
             REDIS.set('lang_cache_{}'.format(chat_id), user_lang['user_lang'])
             return user_lang['user_lang']
@@ -122,8 +135,12 @@ def lang_info(chat_id, pm=False):
     elif locale and pm is True:
         text += "Your locale - `{}`".format(locale)
     buttons = []
-    for a in SUPPORTED_LANGUAGES:
+    for lang in LANGS:
+        print(lang)
+        lang_name = LANGUAGES[lang]["language_info"]["name"]
+        lang_flag = LANGUAGES[lang]["language_info"]["flag"]
+        lang_code = LANGUAGES[lang]["language_info"]["code"]
         buttons.append(
-            [Button.inline(SUPPORTED_LANGUAGES[a],
-             'select_lang_{}'.format(a))])
+            [Button.inline(lang_name + " " + lang_flag,
+             'select_lang_{}'.format(lang_code))])
     return text, buttons
